xe: sdpa: refactor config selection to separate header
